Byg Fix. Dead actors no longer perform actions
The bug was found when I realised the Goblin kept moving after it's
dead. Added If statements to only perform act if the Actor is not dead.
A cleaner way to make sure dead actors don't act would be to remove the
event of the dead actor from the scheduled events.