New Leave Action
New Leave action that allows items carried by the HobbitActor to be
left. Implementing Leave as a HobbitAffordance didn't work since the
take removes the item from the entity manager. Consequently, the
description of the action had to be a much simplier "leave item carried"
instead of "leave"+ the short description of the item carried