Removed render responsibility and dependancies
The render method was completely removed from the Grid as it would now
be handled by the interface (either text or graphical). Consequently the
location width attribute was also removed. This resulted in the removal
of the package dependancy with the Entity manager. Now the HobbitGrid is
much simplier (alot simplier than this description tbh)