GUI 1st refinement
This the previous refinement of the GUI. It will list all commands
together unlike the new GUInterface and will work for any world
regardless of the number of ways in which movements could occur (in
contrast to the new GUInterface that would only work for a 8 way
movement world.) Also a know issue in both GUIs is the message label
refreshes too quickly hence the messages such as "we are on the same
team" when attacking an actor in the same team is missed. This needs to
be fixed. Wow what a long commit message. :)