Chnages to removeEvent method
Changed the way the events queue is updated.

Summary of the findings regarding the bug hunt
Problem
- The Goblin moves one extra step after it has been killed in an attack

Why this happens
- The Goblin’s events are scheduled ahead of the player. Therefore, the
move event of the Goblin is scheduled before the player’s attack. Hence
when the tick happens the move event cause the goblin to move before the
attack which gives the illusion of the goblin moving one extra step
after dying.

What I’ve tried
- Conditions in the tick method of the HobbitActor, to ensure the act
method is not called when an actor is dead
Trying to remove the target’s events from the scheduler if it is killed
in an attack. This didn’t work since the player’s attack is executed
last hence the event priority queue is empty.
- Setting a delay of 1 to the move that is scheduled by the Goblin seems
to solve the problem but it seems like a hack, plus the delay in
unnecessary.

Things that could work
- Scheduling the player’s actions before the anybody elses.