New priority attibute, bug fix and improved Javadoc
New prority attribute was added which enabled Events happening in the
same tick to be ordered according to the priority of the Action of the
Event.  With the addition of the priority attribute Move actions
scheduled in a tick would happen after all the other events, hence the
bug where Goblin moves one extra step after dying was fixed.