Removed the tick and act methods and improved the Javadoc
Tick method is never called hence was removed. The isDead check is
performed in the execute methods of the HobbitAffordance and
HobbitAction to ensure the HobbitActors do not act if they are dead. The
abract act method was removed since it wasn;t overidding the parent
method. Javadoc was improved with the addition of Javadoc to the
attributes.