Monospaced font and highlight human controlled player's location
Human controlled player's locations are highlighted in yellow. The text
of the GUI is mono spaced as requested. New method
locationHasHumanControlledActor returns if or not a given HobbitLocation
has a human controlled actor which is eventually required to colour the
location yellow if it does contain a human controlled actor.