Bug Fix
Fixed the bug where it was possible to chop and attack using a Sword
that had negative hitpoints. The takeDamage method will change the
descriptions of the broken Sword and will remove the CHOPPER and WEAPON
capabilities from the Sword so that a broken Sword cannot be used to
chop or to attack.