Concrete implementation of the GridRenderer Interface
This is the graphical based user interface for the simulation. Is
responsible for outputting a graphical map and messages on a JFrame
window and also obtain user selection of commands from the same JFrame
window.
Its operations are controlled by the HobbitGridController, hence tightly
coupled.

Most of the coding was borrowed from the GUInterface. Coloring of the
locations and human controlled player locations are also performed.