Concrete implementation of GridRenderer Interface
This is the text based user interface for the simulation. Is responsible
for outputting a text based map and messages on the console and
obtaining user selection of commands from the console.

Its operations are controlled by the HobbitGridController hence tightly
coupled.

Most of the coding was borrowed from the TextInterface